requirements.txt
requests~=2.24.0

setuptools~=49.2.0
PyYAML~=5.3.1